pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "datamodel-code-generator>=0.26.5",
    "greenlet>=3.1.1",
    "httpx>=0.28.1",
    "ipython>=8.32.0",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "textual>=1.0.0",
    "textual-dev>=1.7.0",
    "typer>=0.15.1",
    "websockets>=15.0",
]

[dependency-groups]
dev = [
    "pyflowchart>=0.3.1",
]
